API Reference

Jessamine.ACos — Type

Return acos of the operand

Jessamine.ACosAdd — Type

Return acos applied to the sum of the operands

Jessamine.ACosDivide — Type

Return acos applied to the quotient of the operands

Jessamine.ACosMultiply — Type

Return acos applied to the product of the operands

Jessamine.ACosSubtract — Type

Return acos applied to the result of subtraction of the operands

Jessamine.ACosh — Type

Return acosh of the operand

Jessamine.ACoshAdd — Type

Return acosh applied to the sum of the operands

Jessamine.ACoshDivide — Type

Return acosh applied to the quotient of the operands

Jessamine.ACoshMultiply — Type

Return acosh applied to the product of the operands

Jessamine.ACoshSubtract — Type

Return acosh applied to the result of subtraction of the operands

Jessamine.ACot — Type

Return acot of the operand

Jessamine.ACotAdd — Type

Return acot applied to the sum of the operands

Jessamine.ACotDivide — Type

Return acot applied to the quotient of the operands

Jessamine.ACotMultiply — Type

Return acot applied to the product of the operands

Jessamine.ACotSubtract — Type

Return acot applied to the result of subtraction of the operands

Jessamine.ACoth — Type

Return acoth of the operand

Jessamine.ACothAdd — Type

Return acoth applied to the sum of the operands

Jessamine.ACothDivide — Type

Return acoth applied to the quotient of the operands

Jessamine.ACothMultiply — Type

Return acoth applied to the product of the operands

Jessamine.ACothSubtract — Type

Return acoth applied to the result of subtraction of the operands

Jessamine.ACsc — Type

Return acsc of the operand

Jessamine.ACscAdd — Type

Return acsc applied to the sum of the operands

Jessamine.ACscDivide — Type

Return acsc applied to the quotient of the operands

Jessamine.ACscMultiply — Type

Return acsc applied to the product of the operands

Jessamine.ACscSubtract — Type

Return acsc applied to the result of subtraction of the operands

Jessamine.ACsch — Type

Return acsch of the operand

Jessamine.ACschAdd — Type

Return acsch applied to the sum of the operands

Jessamine.ACschDivide — Type

Return acsch applied to the quotient of the operands

Jessamine.ACschMultiply — Type

Return acsch applied to the product of the operands

Jessamine.ACschSubtract — Type

Return acsch applied to the result of subtraction of the operands

Jessamine.ASec — Type

Return asec of the operand

Jessamine.ASecAdd — Type

Return asec applied to the sum of the operands

Jessamine.ASecDivide — Type

Return asec applied to the quotient of the operands

Jessamine.ASecMultiply — Type

Return asec applied to the product of the operands

Jessamine.ASecSubtract — Type

Return asec applied to the result of subtraction of the operands

Jessamine.ASech — Type

Return asech of the operand

Jessamine.ASechAdd — Type

Return asech applied to the sum of the operands

Jessamine.ASechDivide — Type

Return asech applied to the quotient of the operands

Jessamine.ASechMultiply — Type

Return asech applied to the product of the operands

Jessamine.ASechSubtract — Type

Return asech applied to the result of subtraction of the operands

Jessamine.ASin — Type

Return asin of the operand

Jessamine.ASinAdd — Type

Return asin applied to the sum of the operands

Jessamine.ASinDivide — Type

Return asin applied to the quotient of the operands

Jessamine.ASinMultiply — Type

Return asin applied to the product of the operands

Jessamine.ASinSubtract — Type

Return asin applied to the result of subtraction of the operands

Jessamine.ASinh — Type

Return asinh of the operand

Jessamine.ASinhAdd — Type

Return asinh applied to the sum of the operands

Jessamine.ASinhDivide — Type

Return asinh applied to the quotient of the operands

Jessamine.ASinhMultiply — Type

Return asinh applied to the product of the operands

Jessamine.ASinhSubtract — Type

Return asinh applied to the result of subtraction of the operands

Jessamine.ATan — Type

Return atan of the operand

Jessamine.ATanAdd — Type

Return atan applied to the sum of the operands

Jessamine.ATanDivide — Type

Return atan applied to the quotient of the operands

Jessamine.ATanMultiply — Type

Return atan applied to the product of the operands

Jessamine.ATanSubtract — Type

Return atan applied to the result of subtraction of the operands

Jessamine.ATanh — Type

Return atanh of the operand

Jessamine.ATanhAdd — Type

Return atanh applied to the sum of the operands

Jessamine.ATanhDivide — Type

Return atanh applied to the quotient of the operands

Jessamine.ATanhMultiply — Type

Return atanh applied to the product of the operands

Jessamine.ATanhSubtract — Type

Return atanh applied to the result of subtraction of the operands

Jessamine.AbsoluteValue — Type

Return abs of the operand

Jessamine.AbsoluteValueAdd — Type

Return abs applied to the sum of the operands

Jessamine.AbsoluteValueDivide — Type

Return abs applied to the quotient of the operands

Jessamine.AbsoluteValueMultiply — Type

Return abs applied to the product of the operands

Jessamine.AbsoluteValueSubtract — Type

Return abs applied to the result of subtraction of the operands

Jessamine.AbstractGeneOp — Type

An operation to be performed as part of the function of a gene.

Jessamine.AbstractGenome — Type

Abstract base type for genomes.

Jessamine.AbstractLinearModelResult — Type

Abstract base type for the results of fitting linear models.

Jessamine.AbstractModelResult — Type

Abstract base type for the results of fitting a model to data. This object can then be used for predictions.

Jessamine.AbstractPopulationCondition — Type

Base type of conditions for a population. These indicate whether the population is a work in progress, or why its evolution was stopped.

Jessamine.Add — Type

Add operands.

Jessamine.Agent — Type

An agent has a rating, a genome, a parameter vector, and an extra bit of data that depends on how rating is done.

Jessamine.BasicLinearModelResult — Type

A BasicLinearModelResult has a vector of coefficients plust a bias or intercept.

Jessamine.CellState — Type

CellState{EIn,EWork}

Workspace for a genome. The reason for two element type parameters is that this allows EIn to be something like a view into a matrix, while EWork can be a regular Array.

Jessamine.Cos — Type

Return cos of the operand

Jessamine.CosAdd — Type

Return cos applied to the sum of the operands

Jessamine.CosDivide — Type

Return cos applied to the quotient of the operands

Jessamine.CosMultiply — Type

Return cos applied to the product of the operands

Jessamine.CosSubtract — Type

Return cos applied to the result of subtraction of the operands

Jessamine.Cosh — Type

Return cosh of the operand

Jessamine.CoshAdd — Type

Return cosh applied to the sum of the operands

Jessamine.CoshDivide — Type

Return cosh applied to the quotient of the operands

Jessamine.CoshMultiply — Type

Return cosh applied to the product of the operands

Jessamine.CoshSubtract — Type

Return cosh applied to the result of subtraction of the operands

Jessamine.Cot — Type

Return cot of the operand

Jessamine.CotAdd — Type

Return cot applied to the sum of the operands

Jessamine.CotDivide — Type

Return cot applied to the quotient of the operands

Jessamine.CotMultiply — Type

Return cot applied to the product of the operands

Jessamine.CotSubtract — Type

Return cot applied to the result of subtraction of the operands

Jessamine.Coth — Type

Return coth of the operand

Jessamine.CothAdd — Type

Return coth applied to the sum of the operands

Jessamine.CothDivide — Type

Return coth applied to the quotient of the operands

Jessamine.CothMultiply — Type

Return coth applied to the product of the operands

Jessamine.CothSubtract — Type

Return coth applied to the result of subtraction of the operands

Jessamine.Csc — Type

Return csc of the operand

Jessamine.CscAdd — Type

Return csc applied to the sum of the operands

Jessamine.CscDivide — Type

Return csc applied to the quotient of the operands

Jessamine.CscMultiply — Type

Return csc applied to the product of the operands

Jessamine.CscSubtract — Type

Return csc applied to the result of subtraction of the operands

Jessamine.Csch — Type

Return csch of the operand

Jessamine.CschAdd — Type

Return csch applied to the sum of the operands

Jessamine.CschDivide — Type

Return csch applied to the quotient of the operands

Jessamine.CschMultiply — Type

Return csch applied to the product of the operands

Jessamine.CschSubtract — Type

Return csch applied to the result of subtraction of the operands

Jessamine.Divide — Type

Divide operands, as in x[1] / (x[2] * x[3] * ...).

Jessamine.EvolutionSpec — Type

Parameters needed to run random_initial_population and evolution_loop.

Jessamine.Exp — Type

Return exp of the operand

Jessamine.ExpAdd — Type

Return exp applied to the sum of the operands

Jessamine.ExpDivide — Type

Return exp applied to the quotient of the operands

Jessamine.ExpMultiply — Type

Return exp applied to the product of the operands

Jessamine.ExpSubtract — Type

Return exp applied to the result of subtraction of the operands

Jessamine.FzAnd — Type

Return fuzzy AND of the operands

Jessamine.FzNand — Type

Return fuzzy NAND of the operands.

Jessamine.FzNor — Type

Return fuzzy NOR of the operands.

Jessamine.FzNot — Type

Return fznot of the operand

Jessamine.FzNotAdd — Type

Return fznot applied to the sum of the operands

Jessamine.FzNotDivide — Type

Return fznot applied to the quotient of the operands

Jessamine.FzNotMultiply — Type

Return fznot applied to the product of the operands

Jessamine.FzNotSubtract — Type

Return fznot applied to the result of subtraction of the operands

Jessamine.FzOr — Type

Return fuzzy OR of the operands.

Jessamine.Genome — Type

A vector of blocks of instructions. During a time step, each instruction is evaluated on the current work space. For each j, the results of the instructions in block j are collected and added, and this sum is used as the value of slot j in the next work space.

Jessamine.GenomeSpec — Type

A collection of parameters specifying the genome architecture and mutation processes.

Jessamine.Identity — Type

Return identity of the operand

Jessamine.IdentityAdd — Type

Return identity applied to the sum of the operands

Jessamine.IdentityDivide — Type

Return identity applied to the quotient of the operands

Jessamine.IdentityMultiply — Type

Return identity applied to the product of the operands

Jessamine.IdentitySubtract — Type

Return identity applied to the result of subtraction of the operands

Jessamine.Log — Type

Return log of the operand

Jessamine.LogAdd — Type

Return log applied to the sum of the operands

Jessamine.LogDivide — Type

Return log applied to the quotient of the operands

Jessamine.LogMultiply — Type

Return log applied to the product of the operands

Jessamine.LogSubtract — Type

Return log applied to the result of subtraction of the operands

Jessamine.Logit — Type

Return logit of the operand

Jessamine.LogitAdd — Type

Return logit applied to the sum of the operands

Jessamine.LogitDivide — Type

Return logit applied to the quotient of the operands

Jessamine.LogitMultiply — Type

Return logit applied to the product of the operands

Jessamine.LogitSubtract — Type

Return logit applied to the result of subtraction of the operands

Jessamine.Maximum — Type

Return the maximum of the operands.

Jessamine.Minimum — Type

Return the minimum of the operands.

Jessamine.Multiply — Type

Multiply operands.

Jessamine.MutationDist — Type

A collection of distribution objects built from a MutationSpec and used to produce random numbers determining mutations.

Jessamine.MutationDist — Method

MutationDist(m_spec::MutationSpec, src_index_max::Int)

Use m_spec to build a MutationDist. Specify that an operand index must be <= src_index_max.

Jessamine.MutationSpec — Type

A collection of parameters specifying the probabilities of various mutations.

Jessamine.Population — Type

A population is a single generation of living Agents.

Jessamine.Power — Type

Compute a power of a real number. Domain-safe if the exponent is greater than zero.

Jessamine.RDDState — Type

State structure corresponding to ongoing iteration of RandomDuplicateDelete

Jessamine.RandomDuplicateDelete — Type

Wrap an iterator source to produce an iterator that will randomly include some elements yielded by the source twice in a row, and some not at all.

Jessamine.ReciprocalAdd — Type

Add operands and return the reciprocal.

Jessamine.ReciprocalMultiply — Type

Multiply operands and return the reciprocal.

Jessamine.ReciprocalSubtract — Type

Subtract operands and return the reciprocal.

Jessamine.Sec — Type

Return sec of the operand

Jessamine.SecAdd — Type

Return sec applied to the sum of the operands

Jessamine.SecDivide — Type

Return sec applied to the quotient of the operands

Jessamine.SecMultiply — Type

Return sec applied to the product of the operands

Jessamine.SecSubtract — Type

Return sec applied to the result of subtraction of the operands

Jessamine.Sech — Type

Return sech of the operand

Jessamine.SechAdd — Type

Return sech applied to the sum of the operands

Jessamine.SechDivide — Type

Return sech applied to the quotient of the operands

Jessamine.SechMultiply — Type

Return sech applied to the product of the operands

Jessamine.SechSubtract — Type

Return sech applied to the result of subtraction of the operands

Jessamine.SelectionDist — Type

Distribution object for tournament selection.

Jessamine.SelectionSpec — Type

Parameters for tournament selection.

Jessamine.Sigmoid — Type

Return sigmoid of the operand

Jessamine.SigmoidAdd — Type

Return sigmoid applied to the sum of the operands

Jessamine.SigmoidDivide — Type

Return sigmoid applied to the quotient of the operands

Jessamine.SigmoidMultiply — Type

Return sigmoid applied to the product of the operands

Jessamine.SigmoidSubtract — Type

Return sigmoid applied to the result of subtraction of the operands

Jessamine.Sign — Type

Return sign of the operand

Jessamine.SignAdd — Type

Return sign applied to the sum of the operands

Jessamine.SignDivide — Type

Return sign applied to the quotient of the operands

Jessamine.SignMultiply — Type

Return sign applied to the product of the operands

Jessamine.SignSubtract — Type

Return sign applied to the result of subtraction of the operands

Jessamine.Sin — Type

Return sin of the operand

Jessamine.SinAdd — Type

Return sin applied to the sum of the operands

Jessamine.SinDivide — Type

Return sin applied to the quotient of the operands

Jessamine.SinMultiply — Type

Return sin applied to the product of the operands

Jessamine.SinSubtract — Type

Return sin applied to the result of subtraction of the operands

Jessamine.Sinh — Type

Return sinh of the operand

Jessamine.SinhAdd — Type

Return sinh applied to the sum of the operands

Jessamine.SinhDivide — Type

Return sinh applied to the quotient of the operands

Jessamine.SinhMultiply — Type

Return sinh applied to the product of the operands

Jessamine.SinhSubtract — Type

Return sinh applied to the result of subtraction of the operands

Jessamine.SoftMax — Type

Apply the soft-max function to the operands

Jessamine.SoftMin — Type

Apply the soft-min function to the operands

Jessamine.Sqrt — Type

Return sqrt of the operand

Jessamine.SqrtAdd — Type

Return sqrt applied to the sum of the operands

Jessamine.SqrtDivide — Type

Return sqrt applied to the quotient of the operands

Jessamine.SqrtMultiply — Type

Return sqrt applied to the product of the operands

Jessamine.SqrtSubtract — Type

Return sqrt applied to the result of subtraction of the operands

Jessamine.Subtract — Type

Subtract operands, as in x[1] - x[2] - x[3]...

Jessamine.Tan — Type

Return tan of the operand

Jessamine.TanAdd — Type

Return tan applied to the sum of the operands

Jessamine.TanDivide — Type

Return tan applied to the quotient of the operands

Jessamine.TanMultiply — Type

Return tan applied to the product of the operands

Jessamine.TanSubtract — Type

Return tan applied to the result of subtraction of the operands

Jessamine.Tanh — Type

Return tanh of the operand

Jessamine.TanhAdd — Type

Return tanh applied to the sum of the operands

Jessamine.TanhDivide — Type

Return tanh applied to the quotient of the operands

Jessamine.TanhMultiply — Type

Return tanh applied to the product of the operands

Jessamine.TanhSubtract — Type

Return tanh applied to the result of subtraction of the operands

Jessamine.UnaryComposition — Type

Compose a multiary operation and apply a unary operation

Base.isless — Method

isless(x::Agent, y::Agent)

Compare the rating field of the two agents. This definition allows a population to be sorted.

Jessamine.build_op_inventory — Method

build_op_inventory(op_names)

Given a container of strings, build the operation inventory for all combinations of the corresponding unary and multiary operations. Return a named tuple with fields

inventory: array of UnaryCompositions

unknown: array of members of op_names that didn't correspond to any known operation

Jessamine.copy_blank — Method

copy_blank(cs::CellState)
Return a copy of `cs` with zero output and scratch vectors.
The parameter and input vectors are shared with `cs`, and so is the index map.

Jessamine.describe_condition — Method

describe_condition(::AbstractPopulationCondition)

Return a short string describing the population condition.

Jessamine.evolution_loop — Method

evolution_loop(
rng::AbstractRNG,
e_spec::EvolutionSpec
pop_init::Population;
kw_args...
)

Advance pop_init and return the new Population.

Calls evolution_loop() with the specifications unpacked from e_spec.

Keyword arguments are passed to the next implementation of evolution_loop().

Jessamine.extend_if_singleton — Method

extend_if_singleton(v::AbstractArray, m::Int)

If v is a singleton, as in v = [x], return [x, x, ...], that is, a vector of m copies of x. Otherwise, assert that length(v) == m and return v. So the result is always a vector of length m.

Jessamine.fancy_interactive_logging — Function

fancy_interactive_logging(level = Logging.Info)

Set the global logger that filters out messages that aren't from the main module or Jessamine, and that prepends the current time to each message. Messages are only logged if they are at or above the specified level.

Jessamine.fznot — Method

fznot(q)

Apply the fuzzy-not function to q.

Jessamine.generation_size — Method

generation_size(s_spec::SelectionSpec)

Return the size of a generation.

Jessamine.get_op_inventory — Function

get_op_inventory(inventory_name)

Look up the pre-built operation inventory with the given name. If no such inventory is known, use PolynomialInventory Return a named tuple with fields

inventory: array of operations

found: true if inventory_name is a known pre-build operation inventory, false otherwise

Jessamine.infinitely_bad — Method

infinitely_bad(sense::SciMLBase.ObjSense)

Return a value that is considered almost infinitely bad in the given optimization sense.

This returns a finite number of the largest possible magnitude because infinite float values Inf and -Inf can't be compared to any other float value using <.

Jessamine.is_domain_safe — Method

is_domain_safe(op::AbstractMultiOp)

Return true if the operation has no domain restrictions. Return false if the operation has domain restrictions, such as division, in which the denominator cannot be zero. The default is to return false.

Jessamine.is_domain_safe — Method

is_domain_safe(op::AbstractUnaryOp)

Return true if the operation has no domain restrictions. Return false if the operation has domain restrictions, such as division, in which the denominator cannot be zero. The default is to return false.

Jessamine.least_squares_ridge — Method

least_squares_ridge(xs, y, lambda, g_spec, genome, parameter)

Compute ridge regression. Assume xs is an array of columns as predictors and y is a column of target values. Apply run_genome_to_last using parameter as the parameter vector and xs as the inputs. Gather a column of 1s and the output columns as a matrix X. The prediction values are y_hat = X * b, where b is a column of (unknown) coefficients. Solve for the b that minimizes norm(y - y_hat)^2 + lambda * norm(b)^2. If all goes well, return (norm(y - y_hat), b). Otherwise return (Inf, nothing).

Jessamine.least_squares_ridge_grow_and_rate — Method

least_squares_ridge_grow_and_rate(xs, y, lambda_b, lambda_p, lambda_op, g_spec, genome, p_init = zeros(...))

Solve for the parameter vector p that minimizes MSE + lambda_b * norm(b)^2 + lambda_p * norm(p)^2, where y_hat and b are found using least_squares_ridge, and MSE is the mean-square error between y and y_hat.

The solver starts with p_init for the initial value of p. If p_init is nothing or not given, a vector of zeros is used.

If all goes well, return an Agent, whose genome is genome, whose parameter is the best p, and whose extra is a BasicLinearModelResult with coefficient vector b. The agent's rating is the objective function's value plus lambda_op R where R is the total number of operands across all instructions in genome.

Otherwise, return nothing.

Jessamine.logit — Method

logit(t)

Return $\ln\frac{t}{1-t}$. Inverse of sigmoid.

Jessamine.model_predict — Function

model_predict(mr::AbstractModelResult, X; kw_args...)

Given the table X where the columns are inputs (predictors) and the rows are points, use mr to predict the target value for all the points.

Jessamine.model_predict — Method

model_predict(lmr::AbstractLinearModelResult, X::AbstractMatrix)

Return X * coefficients(lmr) + intercept(lmr)

Jessamine.model_predict — Method

model_predict(mr::AbstractLinearModelResult, xs::AbstractArray{<:AbstractArray}; kw_args...)

Stack the columns xs into a matrix X and call model_predict. The kw_args are splatted in.

Jessamine.model_predict — Method

model_predict(g_spec::GenomeSpec, agent::Agent, xs; kw_args...)

Run agent.genome on inputs xs and agent.parameter, and form the linear combination of the genome's outputs using the coefficients agent.extra.

Jessamine.mutate — Function

mutate(rng::AbstractRNG, m_dist::MutationDist, x)

Randomly change x, using random numbers supplied by RNG and the distributions specified by m_dist.

Jessamine.mutation_spec_auto_weight — Function

mutation_spec_auto_weight(op_inventory, weight_scheme; kwargs...)

Build a MutationSpec using the given operation inventory and weight scheme. Other keyword arguments are passed to the constructor of MutationSpec.

Jessamine.new_genome — Method

new_genome(rng::AbstractRNG, s_dist::SelectionDist, m_dist::MutationDist, pop::Population)

Produce a new offspring genome. Pick two parents using tournament selection. Recombine their genomes, and apply mutations.

Jessamine.next_generation — Method

next_generation(
    rng::AbstractRNG,
    g_spec::GenomeSpec,
    s_dist::SelectionDist,
    m_dist::MutationDist,
    pop::Population,
    grow_and_rate;
    valid_agent_max_attempts = 100,
    sense = MinSense)

Produce the next generation of a population by selection and mutation. The offsprings' genomes are produced by new_genome, and fed to grow_and_rate(rng, g_spec, genome), which should "grow" each organism and return the rating and extra data as an Agent, which is inserted into the population. The sense parameter specifies whether selection should minimize or maximize the rating.

Since growing an agent from a genome may fail, the function will try to produce a valid agent up to valid_agent_max_attempts times. After that many failures, it will give up and throw an error.

The sense parameter specifies whether selection should aim to minimize (MinSense) or maximize (MaxSense) the rating of the agents.

Jessamine.num_instructions — Method

num_instructions(genome::Genome)

Return the total number of instructions in all blocks in the genome.

Jessamine.num_operands — Method

num_operands(xs::AbstractArray)

Return the total number of instruction operands in xs.

Jessamine.num_operands — Method

num_operands(genome::Genome)

Return the total number of operands in all instructions in the genome.

Jessamine.num_operands — Method

num_operands(instruction::Instruction)

Return the total number of operands in the instruction.

Jessamine.op_eval — Function

op_eval(op::AbstractGeneOp, operands::AbstractArray)

Return op applied to a list of operands.

Jessamine.op_eval — Method

op_eval(op::Add, workspace, indices)

Evaluate the Add operation on the workspace at the given indices. This function adds the elements in workspace at the specified indices.

Jessamine.op_eval — Method

op_eval(op::Divide, workspace, indices)

Evaluate the Divide operation on the workspace at the given indices.

This function divides the element in workspace at the first index by the product of the elements in workspace at the remaining indices. If the list of indices is empty, the result is 1. If the list of indices has only one index, the result is the workspace element at that index.

Jessamine.op_eval — Method

op_eval(op::FzAnd, workspace, indices)

Evaluate the fuzzy AND operation on the workspace at the given indices.

Jessamine.op_eval — Method

op_eval(op::FzNand, workspace, indices)

Evaluate the fuzzy NAND operation on the workspace at the given indices.

Jessamine.op_eval — Method

op_eval(op::FzNor, workspace, indices)

Evaluate the fuzzy NOR operation on the `workspace` at the given `indices`.

Jessamine.op_eval — Method

op_eval(op::FzOr, workspace, indices)

Evaluate the fuzzy OR operation on the workspace at the given indices.

Jessamine.op_eval — Method

op_eval(op::Multiply, workspace, indices)

Evaluate the Multiply operation on the workspace at the given indices. This function multiplies the elements in workspace at the specified indices.

Jessamine.op_eval — Method

op_eval(op::Subtract, workspace, indices)

Evaluate the Subtract operation on the workspace at the given indices. This function returns 0 if the list of indices is empty, and otherwise computes the result of workspace[indices[1]] - workspace[indices[2]] - ....

Jessamine.op_eval_add_into! — Method

op_eval_add_into!(dest, op, workspace, indices)

Apply the operator op to operands = workspace[indices] and add the result elementwise in place to dest. The default is to do dest .= dest .+ op_eval(op, workspace, indices).

The idea here is that in-place accumulation instead of the default implementation can sometimes avoid the allocation of a temporary array to hold the result of op_eval(...).

Jessamine.pick_parent — Method

pick_parent(rng::AbstractRNG, s_dist::SelectionDist, pop::Population)::Agent

Using tournament selection, pick a parent from the population.

Jessamine.random_genome — Method

random_genome(rng::AbstractRNG, g_spec::GenomeSpec, m_spec::MutationSpec, arity_dist::Distribution; domain_safe=true, restrict_operands=true)

Produce a random genome. It will contain an instruction block for each mutable slot in the state array as specified by g_spec. Each block will contain one random instruction. The operator is picked uniformly at random from m_spec.op_inventory. The number of operands is drawn from arity_dist. The operands are drawn uniformly from the set of possible indices.

If domain_safe is true, only operators for which is_domain_safe(op) is true are used.

If restrict_operands is true, all randomly selected operand indices will come from parameters and inputs.

Jessamine.random_initial_population — Method

random_initial_population(
    rng::AbstractRNG,
    e_spec::EvolutionSpec,
    arity_dist::Distribution;
    domain_safe = true,
    restrict_operands = true,
    sense = MinSense)

Initialize a random initial population from an e_spec.

Jessamine.random_initial_population — Method

random_initial_population(
    rng::AbstractRNG,
    g_spec::GenomeSpec,
    m_dist::MutationDist,
    arity_dist::Distribution,
    s_spec::SelectionSpec,
    grow_and_rate;
    domain_safe = true,
    restrict_operands = true,
    valid_agent_max_attempts = 100,
    sense = MinSense)::Population

Make a random initial population. The number of genomes is specified by adding the number of new genomes per generation and the number of keepers given in s_spec. Other parameters are passed to random_genome to produce random genomes. The grow_and_rate function is called with rng, g_spec, genome.

Since growing an agent from a genome may fail, the function will try to produce a valid agent up to valid_agent_max_attempts times. After that many failures, it will give up and throw an error.

The domain_safe and restrict_operands parameters are passed to random_genome.

The sense parameter specifies whether selection should aim to minimize (MinSense) or maximize (MaxSense) the rating of the agents.

Jessamine.run_genome — Method

run_genome(g_spec::GenomeSpec, genome::Genome, parameter::AbstractArray, input)::Vector

Build a work space vector using zeros for each output and scratch slot, followed by the parameter vector, then the input vector. Evaluate the instructions in genome, repeating the evaluation g_spec.num_time_steps. Return an array that contains, for each time step, the elements 1 through g_spec.output_size of the work space vector.

Jessamine.run_genome_time_series — Method

TODO add documentation

Jessamine.run_genome_to_last — Method

run_genome_to_last(g_spec::GenomeSpec, genome::AbstractGenome, parameter::AbstractArray, input)::Vector

Run the genome on the given inputs with the given parameter vector. Return the vector of ouptuts from the final time step. This is the most general implementation.

Jessamine.short_show — Method

short_show(x)

Return a short string representation of x.

Jessamine.short_show — Method

short_show(io::IO, x)

Print a short version of x to io

Jessamine.sigmoid — Method

sigmoid(t)

Return $\frac{1}{1 + \exp(-t)}$. Inverse of logit.

Jessamine.splat_or_default — Method

splat_or_default(op, def, workspace, indices)

Return the result of applying op to the operands = workspace[indices], with op([]) = def and op([x1...]) = op(x1, op(x2, ...)). The Base.splat function doesn't have a way to deal with the first case. The operation should be flat (commutative and associative).

Jessamine.time_series_least_squares_ridge — Method

time_series_least_squares_ridge(xs, y, lambda, g_spec, genome, parameter; memory_steps = 1)

Compute ridge regression. Assume xs is a vector of state points and y is a column of target values. Apply run_genome_time_series using parameter as the parameter vector and xs as the inputs. Pass through the given value of memory_steps. Gather a column of 1s and the output columns as a matrix X. The prediction values are y_hat = X * b, where b is a column of (unknown) coefficients. Solve for the b that minimizes norm(y - y_hat)^2 + lambda * norm(b)^2. If all goes well, return (norm(y - y_hat), b). Otherwise return (Inf, nothing).

Jessamine.time_series_least_squares_ridge_grow_and_rate — Method

time_series_least_squares_ridge_grow_and_rate(xs, y, lambda_b, lambda_p, lambda_op, g_spec, genome, memory_steps = 1, p_init = zeros(...))

Solve for the parameter vector p that minimizes norm(y - y_hat)^2 + lambda_b * norm(b)^2 + lambda_p * norm(p)^2 + lambda_op R, where y_hat and b are found using time_series_least_squares_ridge. R is the total number of operands across all instructions in genome.

The solver starts with p_init for the initial value of p. If p_init is nothing or not given, a vector of zeros is used.

If all goes well, return an Agent, whose genome is genome, whose parameter is the best p, and whose extra is a BasicLinearModelResult with coefficient vector b.

Otherwise, return nothing.

Jessamine.to_expr — Function

to_expr(op, cell_state, operands)

Given an AbstractGeneOp, a cell state, and a vector of operands of the form (field, index), build a Julia expression that represents the application of op to the operands cell_state.field1[operand1], cell_state.field2[operand2], ... This function is for the implementation of compiled genomes.

Jessamine.v_convert — Method

v_convert(ref::Type, x)

Just convert x to the reference type.

Jessamine.v_convert — Method

v_convert(ref::Type{<:AbstractArray}, x::AbstractArray)

Convert an array-like object x to a regular Array. This function is used to create arrays within run_genome_* functions.

This function can be specialized for array types for which another array type is more appropriate.

Jessamine.v_convert — Method

v_convert(ref::Type{<:AbstractArray}, x)

Convert an x that is not an array to an array of the reference type with just one element equal to x.

Jessamine.very_short_show — Method

very_short_show(x)

Return a very short string representation of x.

Jessamine.very_short_show — Method

very_short_show(io::IO, x)

Print a very short version of x to io

Jessamine.vns_evolution_loop — Method

vns_evolution_loop(
rng::AbstractRNG,
neighborhoods::AbstractArray{EvolutionSpec},
pop_init::Population;
kw_args...
)

Perform variable neighborhood search.

Evolution is performed using neighborhoods[1] using evolution_loop. That's the first epoch. If a new best rating is achieved, continue with that neighborhood in the second epoch. Otherwise, switch to using neighborhoods[2] etc. Once a new best rating is achieved, return to neighborhoods[1].

Fields from the neighborhoods and relevant keyword arguments are passed as keyword arguments to evolution_loop.

Other keyword arguments:

verbosity = 1:

Setting verbosity to 0 emits @debug messages instead of @info messages. Log messages are produced after each epoch completes.

max_epochs = nothing:

If max_epochs is an integer rather than nothing, the loop ends after at most that many epochs.

discovery_channel = nothing:

If discovery_channel is a Channel rather than nothing, each time the best agent in a generation is better than all previous ones, it is sent as a message to this channel.

stop_threshold = nothing:

If stop_threshold is a number rather than nothing, the loop stops as soon as an agent with a rating better than stop_threshold is discovered.

stop_channel = nothing:

If stop_channel is a Channel rather than nothing, the loop checks for a message on stop_channel before calling next_generation. If a message is found, the loop stops.

stop_deadline = nothing:

If stop_deadline is a DateTime rather than nothing, the loop checks whether now() > stop_deadline before calling next_generation, and if so, the loop stops.

valid_agent_max_attempts = 100:

Passed to next_generation().

Jessamine.vns_keep_going — Method

vns_keep_going(::AbstractPopulationCondition)

Return whether VNS should continue to another epoch. Utility function.

Jessamine.weight — Function

weight(op, weight_scheme)

Return a positive number that indicates the weight of the operation. During mutation, operations are selected at rates proportional to the reciprocal of their weights, so heavier operations are less likely to be selected.

Jessamine.workspace_size — Method

workspace_size(g_spec::GenomeSpec)

Return the number of elements in the workspace vector specified by g_spec.

Jessamine.@debug_or_info — Macro

@debug_or_info(verbosity, messages...)

If the expression verbosity evaluates to a number greater than 0, call @info with messsges, otherwise call @debug with messages.

That is, the messges are used to create a @debug log item, unless verbosity is turned on, in which case it's promoted to a @info log item.