API Reference

Jessamine.AbstractGeneOp — Type

An operation to be performed as part of the function of a gene.

Jessamine.AbstractGenome — Type

Abstract base type for genomes.

Jessamine.AbstractLinearModelResult — Type

Abstract base type for the results of fitting linear models.

Jessamine.AbstractMachineSpec — Type

Abstract base type for machine specs

Jessamine.AbstractModelResult — Type

Abstract base type for the results of fitting a model to data. This object can then be used for predictions.

Jessamine.AbstractPopulationCondition — Type

Base type of conditions for a population. These indicate whether the population is a work in progress, or why its evolution was stopped.

Jessamine.AbstractSolverSpec — Type

Abstract base type for solver specs, used for solving for genome parameter values.

Jessamine.Add — Type

Add operands.

Jessamine.Agent — Type

An agent has a rating, a genome, a parameter vector, and an extra bit of data that depends on how rating is done.

Jessamine.BasicLinearModelResult — Type

A BasicLinearModelResult has a vector of coefficients plust a bias or intercept.

Jessamine.DefaultSolverSpec — Type

Use this if you want all default solving procedures.

Jessamine.EvolutionSpec — Type

Parameters needed to run random_initial_population and evolution_loop.

Jessamine.FzAnd — Type

Return fuzzy AND of the operands

Jessamine.FzNand — Type

Return fuzzy NAND of the operands.

Jessamine.FzNor — Type

Return fuzzy NOR of the operands.

Jessamine.FzOr — Type

Return fuzzy OR of the operands.

Jessamine.Genome — Type

A vector of blocks of instructions. During a time step, each instruction is evaluated on the current work space. For each j, the results of the instructions in block j are collected and added, and this sum is used as the value of slot j in the next work space.

Jessamine.GenomeSpec — Type

A collection of parameters specifying the genome architecture and mutation processes.

Jessamine.LinearModelMachineSpec — Type

Wrap an MLJ Model that uses a linear combination of input columns to produce its predictions.

Jessamine.LinearModelMachineSpec — Method

LinearModelMachineSpec(model::Model, lambda_parameter, lambda_operand)

Construct a LinearModelMachineSpec using model.lambda for lambda_model.

Jessamine.Maximum — Type

Return the maximum of the operands.

Jessamine.Minimum — Type

Return the minimum of the operands.

Jessamine.Multiply — Type

Multiply operands.

Jessamine.MutationDist — Type

A collection of distribution objects built from a MutationSpec and used to produce random numbers determining mutations.

Jessamine.MutationDist — Method

MutationDist(m_spec::MutationSpec, src_index_max::Int)

Use m_spec to build a MutationDist. Specify that an operand index must be <= src_index_max.

Jessamine.MutationSpec — Type

A collection of parameters specifying the probabilities of various mutations.

Jessamine.Population — Type

A population is a single generation of living Agents.

Jessamine.RDDState — Type

State structure corresponding to ongoing iteration of RandomDuplicateDelete

Jessamine.RandomDuplicateDelete — Type

Wrap an iterator source to produce an iterator that will randomly include some elements yielded by the source twice in a row, and some not at all.

Jessamine.ReciprocalAdd — Type

Add operands and return the reciprocal.

Jessamine.ReciprocalMultiply — Type

Multiply operands and return the reciprocal.

Jessamine.ReciprocalSubtract — Type

Subtract operands and return the reciprocal.

Jessamine.SelectionDist — Type

Distribution object for tournament selection.

Jessamine.SelectionSpec — Type

Parameters for tournament selection.

Jessamine.SignAdd — Type

Return the sign of the sum of the operands

Jessamine.Subtract — Type

Subtract LISP-style: 0 + x[1] - x[2] - x[3]...

Jessamine.UnaryComposition — Type

Do a multi-arity operation and apply a unary operation

Base.isless — Method

isless(x::Agent, y::Agent)

Compare the rating field of the two agents. This definition allows a population to be sorted.

Jessamine.area_above_curve — Method

area_above_curve(y_hat, y_ref)

Return the area above the ROC curve for the predictions y_hat and actual reference values y_ref. This is 1 minus the area under the curve. The vector y_hat should consist of Bernoulli distributions, as returned by a LogisticClassifier, for example. The vector y should consist of values from those distributions.

Jessamine.describe_condition — Method

describe_condition(::AbstractPopulationCondition)

Return a short string describing the population condition.

Jessamine.do_simplification — Method

do_simplification(::AbstractPopulationCondition)

Return whether to perform the simplification epoch. Utility function.

Jessamine.eval_time_step_symbolic — Method

eval_time_step_symbolic(g_spec, genome; output_sym, scratch_sym, parameter_sym, input_sym)

Return symbolic objects representing a single time step of genome. The result is a named tuple with fields z, t, p, x for the Symbolics objects for those variables; c and c_next for the current and future cell state in symbolic form.

Jessamine.extend_if_singleton — Method

extend_if_singleton(v::AbstractVector, m::Int)

If v is a singleton, as in v = [x], return [x, x, ...], that is, a vector of m copies of x. Otherwise, assert that length(v) == m and return v. So the result is always a vector of length m.

Jessamine.generation_size — Method

generation_size(s_spec::SelectionSpec)

Return the size of a generation.

Jessamine.genome_complexity — Method

genome_complexity(mn_spec, g_spec, genome)

Numerical complexity of genome. The default implementation returns mn_spec.lambda_operand times the number of operands in the genome.

Jessamine.genome_parameter_complexity — Method

genome_parameter_complexity(mn_spec, p)

Numerical complexity of a parameter vector. The default implementation returns mn_spec.lambda_parameter times the squared L2 norm of p.

Jessamine.least_squares_ridge — Method

least_squares_ridge(xs, y, lambda, g_spec, genome, parameter)

Compute ridge regression. Assume xs is an array of columns as predictors and y is a column of target values. Apply run_genome using parameter as the parameter vector and xs as the inputs. Gather a column of 1s and the output columns as a matrix X. The prediction values are y_hat = X * b, where b is a column of (unknown) coefficients. Solve for the b that minimizes norm(y - y_hat)^2 + lambda * norm(b)^2. If all goes well, return (norm(y - y_hat), b). Otherwise return nothing.

Jessamine.least_squares_ridge_grow_and_rate — Function

least_squares_ridge_grow_and_rate(xs, y, lambda_b, lambda_p, lambda_op, g_spec, genome, p_init = zeros(...))

Solve for the parameter vector p that minimzes norm(y - y_hat)^2 + lambda_b * norm(b)^2 + lambda_p * norm(p)^2 + lambda_op R, where y_hat and b are found using least_squares_ridge. R is the total number of operands across all instructions in genome.

The solver starts with p_init for the initial value of p. If p_init is nothing or not given, a vector of zeros is used.

If all goes well, return an Agent, whose genome is genome, whose parameter is the best p, and whose extra is a BasicLinearModelResult with coefficient vector b.

Otherwise, return nothing.

Jessamine.machine_complexity — Method

machine_complexity(mn_spec, m; kw_args...)

Numerical complexity of a machine.

The default is to return 0.

Jessamine.machine_complexity — Method

machine_complexity(mn_spec, m)

Numerical complexity of a machine.

Return the sum of squares of MLJ.fitted_params(m) multiplied by mn_spec.lambda_model.

Jessamine.machine_fit! — Method

machine_fit!(mn_spec, m; kw_args...)

The default implementation returns MLJ.fit!(m; kw_args...).

Jessamine.machine_grow_and_rate — Function

machine_grow_and_rate(xs, y, g_spec, genome, mn_spec, sol_spec, w = nothing)

Solve for the parameter values that produce the best machine for predicting y from the outputs of the genome applied to xs. A non-nothing value of w is passed to the machine as a weight vector. The resulting parameter vector is stored as the parameter field in the returned Agent. The resulting machine is wrapped in a MachineResult and stored in the extra field of the Agent.

If any exception is thrown during the solving process, the exception is suppressed, and nothing is returned.

Jessamine.machine_init — Method

machine_init(mn_spec, X, y; kw_args...)

Construct a machine with predictor table X and target y.

The default implementation calls MLJ.machine(mn_spec.machine, X, y; kw_args...)

Jessamine.machine_predict — Method

machine_predict(mn_spec, m, X; kw_args...)

Produce a prediction y_hat. The default implementation returns MLJ.predict(m, X; kw_args...).

Jessamine.machine_spec_type — Method

machine_spec_type(::Type{LassoRegressor})

Return LinearModelMachineSpec

Jessamine.machine_spec_type — Method

machine_spec_type(::Type{LogisticClassifier})

Return LinearModelMachineSpec

Jessamine.machine_spec_type — Method

machine_spec_type(::Type{RidgeRegressor})

Return LinearModelMachineSpec

Jessamine.machine_spec_type — Method

machine_spec_type(t_MLJ_model::Type)::Type{<:AbstractMachineSpec}

Return the Jessamine type (subtype of AbstractMachineSpec) corresponding to the given MLJ model type. The default is to return BasicModelMachineSpec

Jessamine.model_predict — Function

model_predict(mr::AbstractModelResult, X; kw_args...)

Given the table X where the columns are inputs (predictors) and the rows are points, use mr to predict the target value for all the points.

Jessamine.model_predict — Method

model_predict(mr::AbstractLinearModelResult, xs::AbstractVector{<:AbstractVector}; kw_args...)

Stack the columns xs into a matrix X and call model_predict. The kw_args are splatted in.

Jessamine.model_predict — Method

model_predict(lmr::AbstractLinearModelResult, X::Matrix)

Return X * coefficients(lmr) + intercept(lmr)

Jessamine.model_predict — Method

model_predict(g_spec::GenomeSpec, agent::Agent, xs; kw_args...)

Run agent.genome on inputs xs and agent.parameter, and form the linear combination of the genome's outputs using the coefficients agent.extra.

Jessamine.model_symbolic_output — Method

model_symbolic_output(g_spec, agent; kw_args...)

Build a Symbolics form for the output of the final time step of running agent's genome Then use the agent's parameter vector, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form.

The kw_args are eventually splatted into model_predict.

Jessamine.model_symbolic_output — Method

model_symbolic_output(g_spec, genome, parameter, model_result; kw_args...)

Build a Symbolics form for the output of the final time step of running genome Then use the parameter vector, the symbolic output of the genome, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form.

The kw_args are eventually splatted into model_predict.

Return a named tuple with lots of useful fields.

Jessamine.model_symbolic_output — Method

model_symbolic_output(fit_result::NamedTuple; kw_args...)

Call model_symbolic_output with the g_spec::GenomeSpec and best_agent::Agent found in fit_result. Return the result.

Jessamine.model_sympy_output — Method

model_sympy_output(g_spec, agent; kw_args...)

Build a SymPy form for the output of the final time step of running agent's genome Then use the agent's parameter vector, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form.

Key-word arguments:

assumptions = Dict(:extended_real => true):

Assumptions to use when creating symbols.

Other kw_args are eventually splatted into model_predict.

Jessamine.model_sympy_output — Method

model_sympy_output(g_spec, genome, parameter, model_result; kw_args...)

Build a SymPy form for the output of the final time step of running genome Then use the parameter vector, the symbolic output of the genome, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form.

Key-word arguments:

assumptions = Dict(:extended_real => true):

Assumptions to use when creating symbols.

Other kw_args are eventually splatted into model_predict.

Return a named tuple with lots of useful fields.

Jessamine.mutate — Function

mutate(rng::AbstractRNG, m_dist::MutationDist, x)

Randomly change x, using random numbers supplied by RNG and the distributions specified by m_dist.

Jessamine.new_genome — Method

new_genome(rng::AbstractRNG, s_dist::SelectionDist, m_dist::MutationDist, pop::Population)

Produce a new offspring genome. Pick two parents using tournament selection. Recombine their genomes, and apply mutations.

Jessamine.next_generation — Method

next_generation(
    rng::AbstractRNG,
    g_spec::GenomeSpec,
    s_dist::SelectionDist,
    m_dist::MutationDist,
    pop::Population,
    grow_and_rate;
    sense = MinSense)

Produce the next generation of a population by selection and mutation. The offsprings' genomes are produced by new_genome, and fed to grow_and_rate(rng, g_spec, genome), which should "grow" each organism and return the rating and extra data as an Agent, which is inserted into the population. The sense parameter specifies whether selection should minimize or maximize the rating.

Jessamine.num_instructions — Method

num_instructions(genome::Genome)

Return the total number of instructions in all blocks in the genome.

Jessamine.num_operands — Method

num_operands(xs::AbstractVector)

Return the total number of instruction operands in xs.

Jessamine.num_operands — Method

num_operands(genome::Genome)

Return the total number of operands in all instructions in the genome.

Jessamine.num_operands — Method

num_operands(instruction::Instruction)

Return the total number of operands in the instruction.

Jessamine.op_eval — Function

op_eval(op::AbstractGeneOp, operands::AbstractVector)

Return op applied to a list of operands.

Jessamine.op_eval_add_into! — Method

op_eval_add_into!(dest, op, operands)

Apply the operator op to operands and add the result elementwise in place to dest. The default is to do dest .= dest .+ op_eval(op, operands).

The idea here is that in-place accumulation instead of the default implementation can sometimes avoid the allocation of a temporary array to hold the result of op_eval(...).

Jessamine.parameter_solver_optimization_function — Method

parameter_solver_optimization_function(sol_spec, f)

Build an OptimizationFunction around the function f(genome_parameter_vector, _). The resulting objective function object will eventually be passed to solve(). The default implementation returns OptimizationFunction(f).

Jessamine.parameter_solver_optimization_problem — Method

parameter_solver_optimization_problem(sol_spec, g_spec, optim_fn, context)

Build an OptimizationProblem around the OptimizationFunction. The resulting problem object will be passed to solve. The default implementation returns OptimizationProblem(optim_fn, zeros(...), context, sense=MinSense).

Jessamine.parameter_solver_solve — Method

parameter_solver_solve(sol_spec, optim_prob)

Solve the problem generated by parameter_solver_optimization_problem(). The default implementation returns solve(optim_prob, NelderMeade())

Jessamine.pick_parent — Method

pick_parent(rng::AbstractRNG, s_dist::SelectionDist, pop::Population)::Agent

Using tournament selection, pick a parent from the population.

Jessamine.prediction_performance — Method

prediction_performance(mn_spec, y_hat, y_ref)

Return a performance measure of the prediction y_hat compared to reference y_ref.

The default implementation applies the callable mn_spec.performance(y_hat, y_ref)

Jessamine.random_genome — Method

random_genome(rng::AbstractRNG, g_spec::GenomeSpec, m_spec::MutationSpec, arity_dist::Distribution)

Produce a random genome. It will contain an instruction block for each mutable slot in the state array as specified by g_spec. Each block will contain one random instruction. The operator is picked uniformly at random from m_spec.op_inventory. The number of operands is drawn from arity_dist. The operands are drawn uniformly from the set of possible indices.

Jessamine.random_initial_population — Method

random_initial_population(
    rng::AbstractRNG,
    e_spec::EvolutionSpec,
    arity_dist::Distribution;
    sense = MinSense)

Initialize a random initial population from an e_spec.

Jessamine.random_initial_population — Method

random_initial_population(
    rng::AbstractRNG,
    g_spec::GenomeSpec,
    m_dist::MutationDist,
    arity_dist::Distribution,
    s_spec::SelectionSpec,
    grow_and_rate;
    sense = MinSense)::Population

Make a random initial population. The number of genomes is specified by adding the number of new genomes per generation and the number of keepers given in s_spec. Other parameters are passed to random_genome to produce random genomes. The grow_and_rate function is called with rng, g_spec, genome.

Jessamine.replace_near_integer — Method

replace_near_integer(expr; tolerance=1.0e-10)

Round literal floating-point numbers within a symbolic expression. Specifically, if a number a differs round(a) by less thant tolerance, it gets replaced by round(a).

Jessamine.run_genome — Method

run_genome(g_spec::GenomeSpec, genome::Genome, parameter::AbstractVector, input)::Vector

Build a work space vector using zeros for each output and scratch slot, followed by the parameter vector, then the input vector. Evaluate the instructions in genome, repeating the evaluation g_spec.num_time_steps. Return an array that contains, for each time step, the elements 1 through g_spec.output_size of the work space vector.

Jessamine.run_genome_symbolic — Method

run_genome_symbolic(g_spec, genome; paramter_sym=:p, input_sym=:x)

Build a symbolic form for the output of the final time step of running genome. The parameter vector and input vector are Symbolics objects of the form p[j] and x[j]. The variable names can be specified with the keyword arguments.

Return a named tuple (p, x, z) where p and x, are vectors of Symbolics objects used to represent genome parameters and inputs; and w is a vector of genome outputs in symbolic form.

Jessamine.short_show — Method

short_show([io::IO], x)

Print a short version of x to io, using stdout by default.

Jessamine.show_symbolic — Method

show_symbolic(g_spec, genome; output_sym, scratch_sym, parameter_sym, input_sym)

Return a matrix with three columns. The first is row indices 1, 2, etc. The second is the inital worspace vector in symbolic form. The third is the future workspace state in symbolic form.

Jessamine.show_symbolic — Method

show_symbolic(fit_result::NamedTuple; kw_args...)

Call show_symbolic with the g_spec::GenomeSpec and best_agent.genome::AbstractGenome found in fit_result. Return the result.

Jessamine.splat_or_default — Method

splat_or_default(op, def, operands)

Return the result of applying op to the operands, with op([]) = def and op([x1...]) = op(x1...). The Base.splat function doesn't have a way to deal with the first case.

Jessamine.vns_keep_going — Method

vns_keep_going(::AbstractPopulationCondition)

Return whether VNS should continue to another epoch. Utility function.

Jessamine.workspace_size — Method

workspace_size(g_spec::GenomeSpec)

Return the number of elements in the workspace vector specified by g_spec.

StatisticalTraits.supports_weights — Method

supports_weights(::Type{JessamineModel})

Return true. A Jessamine model supports weight vectors for training data if the inner model does. Since MLJ determines this at the type level, it has to return true here to allow a weight vector to be passed in. The fit implementation throws an exception if a non-nothing weight vector is passed in and the inner model does not allow a weight vector.

StatisticalTraits.supports_weights — Method

supports_weights(::Type{JessamineModel})

Return true. A Jessamine model supports weight vectors for training data if the inner model does. Since MLJ determines this at the type level, it has to return true here to allow a weight vector to be passed in. The fit implementation throws an exception if a non-nothing weight vector is passed in and the inner model does not allow a weight vector.